+++ bathroom_app/lib/Dashboard/upload_second.dart
@@ -1,211 +1,206 @@
+import 'package:bathroom_app/Dashboard/upload_third.dart';
 import 'package:flutter/material.dart';
+import 'package:image_picker/image_picker.dart';
+import 'package:bathroom_app/Dashboard/image_added.dart';
 import 'package:bathroom_app/Dashboard/tags.dart';
-import 'package:bathroom_app/Dashboard/upload_success.dart';
 
 class SecondUploadPage extends StatefulWidget {
   const SecondUploadPage({Key? key}) : super(key: key);
 
   @override
-  State<SecondUploadPage> createState() => SecondUploadPageState();
+  State<SecondUploadPage> createState() => _SecondUploadPageState();
 }
 
-class SecondUploadPageState extends State<SecondUploadPage> {
-  int rating = 2;
-  final TextEditingController reviewController = TextEditingController();
+class _SecondUploadPageState extends State<SecondUploadPage> {
+  final List<XFile> _images = [];
+  final ImagePicker _picker = ImagePicker();
 
-  final List<String> tags = [
-    'Test',
-    'Bidet',
-    'Soap',
-    'Clean',
-    'Dog water',
-    'LGBT-friendly',
-  ];
-
-  @override
-  void dispose() {
-    reviewController.dispose();
-    super.dispose();
+  Future<void> _pickImages() async {
+    try {
+      final XFile? picked = await _picker.pickImage(
+        source: ImageSource.gallery,
+      );
+      if (picked != null) {
+        setState(() {
+          _images.add(picked);
+        });
+      }
+    } catch (_) {}
   }
 
-  Widget buildStar(int index) {
-    bool filled = index < rating;
-    return GestureDetector(
-      onTap: () {
-        setState(() {
-          rating = index + 1;
-        });
-      },
-      child: Icon(
-        Icons.star,
-        color: filled ? Colors.orange : Colors.grey,
-        size: 32,
-      ),
-    );
+  void _removeImage(int index) {
+    setState(() {
+      _images.removeAt(index);
+    });
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: Colors.white,
-      body: SafeArea(
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 24),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              // Progress bar row, centered
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
+      body: Stack(
+        children: [
+          // Page content
+          SafeArea(
+            child: Padding(
+              padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 24),
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
-                  Container(
-                    width: 40,
-                    height: 6,
-                    decoration: BoxDecoration(
-                      color: Colors.orange,
-                      borderRadius: BorderRadius.circular(3),
-                    ),
-                  ),
-                  const SizedBox(width: 6),
-                  Container(
-                    width: 40,
-                    height: 6,
-                    decoration: BoxDecoration(
-                      color: Colors.orange,
-                      borderRadius: BorderRadius.circular(3),
-                    ),
-                  ),
-                ],
-              ),
-
-              Row(
-                children: [
-                  GestureDetector(
-                    onTap: () => Navigator.of(context).pop(),
-                    child: Container(
-                      padding: const EdgeInsets.all(8),
-                      decoration: BoxDecoration(
-                        color: Colors.orange,
-                        borderRadius: BorderRadius.circular(24),
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: [
+                      Container(
+                        width: 40,
+                        height: 6,
+                        decoration: BoxDecoration(
+                          color: Colors.orange,
+                          borderRadius: BorderRadius.circular(3),
+                        ),
                       ),
-                      child: const Icon(Icons.arrow_back, color: Colors.white),
+                      const SizedBox(width: 6),
+                      Container(
+                        width: 40,
+                        height: 6,
+                        decoration: BoxDecoration(
+                          color: Colors.orange,
+                          borderRadius: BorderRadius.circular(3),
+                        ),
+                      ),
+                      const SizedBox(width: 6),
+                      Container(
+                        width: 40,
+                        height: 6,
+                        decoration: BoxDecoration(
+                          color: Colors.grey[300],
+                          borderRadius: BorderRadius.circular(3),
+                        ),
+                      ),
+                    ],
+                  ),
+                  Row(
+                    children: [
+                      GestureDetector(
+                        onTap: () => Navigator.of(context).pop(),
+                        child: Container(
+                          padding: const EdgeInsets.all(8),
+                          decoration: BoxDecoration(
+                            color: Colors.orange,
+                            borderRadius: BorderRadius.circular(24),
+                          ),
+                          child: const Icon(
+                            Icons.arrow_back,
+                            color: Colors.white,
+                          ),
+                        ),
+                      ),
+                    ],
+                  ),
+                  const Center(
+                    child: Text(
+                      'Upload',
+                      style: TextStyle(
+                        fontSize: 32,
+                        fontWeight: FontWeight.bold,
+                      ),
                     ),
                   ),
-                ],
-              ),
+                  const SizedBox(height: 24),
 
-              const Align(
-                alignment: Alignment.center,
-                child: Text(
-                  'Review',
-                  style: TextStyle(fontSize: 32, fontWeight: FontWeight.bold),
-                ),
-              ),
-
-              const SizedBox(height: 12),
-
-              // Stars row
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: List.generate(5, (index) => buildStar(index)),
-              ),
-
-              const SizedBox(height: 16),
-
-              // Review input
-              TextField(
-                controller: reviewController,
-                maxLines: 6,
-                decoration: InputDecoration(
-                  hintText: 'Write your review here',
-                  border: OutlineInputBorder(
+                  // Upload area
+                  InkWell(
+                    onTap: _pickImages,
                     borderRadius: BorderRadius.circular(8),
-                  ),
-                  contentPadding: const EdgeInsets.symmetric(
-                    horizontal: 12,
-                    vertical: 12,
-                  ),
-                ),
-              ),
-
-              const SizedBox(height: 24),
-
-              // Tags Title
-              const Align(
-                alignment: Alignment.center,
-                child: Text(
-                  'Tags',
-                  style: TextStyle(fontSize: 32, fontWeight: FontWeight.bold),
-                ),
-              ),
-
-              const SizedBox(height: 12),
-
-              // Tags input/search (no functionality for search)
-              Container(
-                height: 40,
-                padding: const EdgeInsets.symmetric(horizontal: 12),
-                decoration: BoxDecoration(
-                  border: Border.all(color: Colors.grey),
-                  borderRadius: BorderRadius.circular(8),
-                ),
-                child: Row(
-                  children: const [
-                    Icon(Icons.tag, color: Colors.grey),
-                    SizedBox(width: 8),
-                    Expanded(
-                      child: Text(
-                        'Search for tags to add',
-                        style: TextStyle(color: Colors.grey),
+                    child: Container(
+                      width: double.infinity,
+                      height: 140,
+                      decoration: BoxDecoration(
+                        border: Border.all(color: Colors.grey),
+                        borderRadius: BorderRadius.circular(8),
+                        color: Colors.white,
+                      ),
+                      child: Center(
+                        child: Column(
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          children: [
+                            Icon(Icons.add, color: Colors.orange, size: 36),
+                            const SizedBox(height: 8),
+                            Text(
+                              'Add your images here!\nBrowse and add images to be posted',
+                              textAlign: TextAlign.center,
+                              style: TextStyle(
+                                color: Colors.grey[700],
+                                fontWeight: FontWeight.w600,
+                              ),
+                            ),
+                          ],
+                        ),
                       ),
                     ),
-                  ],
-                ),
-              ),
-
-              const SizedBox(height: 12),
-
-              // Tags list - wrap to next line if overflow
-              Expanded(
-                child: SingleChildScrollView(
-                  child: Wrap(
-                    children: tags.map((tag) => TagWidget(text: tag)).toList(),
                   ),
-                ),
-              ),
 
-              // Submit button
-              SizedBox(
-                width: double.infinity,
-                child: ElevatedButton(
-                  onPressed: () {
-                    Navigator.of(context).pushAndRemoveUntil(
-                      MaterialPageRoute(
-                        builder: (_) => const UploadSuccessPage(),
-                      ),
-                      (route) => false,
-                    );
-                  },
-                  style: ElevatedButton.styleFrom(
-                    backgroundColor: Colors.orange,
-                    shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(12),
-                    ),
-                    padding: const EdgeInsets.symmetric(vertical: 14),
-                  ),
-                  child: const Text(
-                    'Submit',
-                    style: TextStyle(
-                      fontWeight: FontWeight.bold,
-                      fontSize: 20,
-                      color: Colors.white,
+                  const SizedBox(height: 16),
+
+                  // Uploaded images
+                  Expanded(
+                    child: ListView.builder(
+                      itemCount: _images.length,
+                      itemBuilder: (context, index) {
+                        final img = _images[index];
+                        return Padding(
+                          padding: const EdgeInsets.only(bottom: 8.0),
+                          child: ImageAddedWidget(
+                            imageName: img.name,
+                            progressPercent: 41,
+                            onRemove: () => _removeImage(index),
+                          ),
+                        );
+                      },
                     ),
                   ),
-                ),
+
+                  const SizedBox(
+                    height: 80,
+                  ), // Leave space for fixed "Next" button
+                ],
               ),
-            ],
+            ),
           ),
-        ),
+
+          // Next Button - Fixed to bottom
+          Positioned(
+            bottom: 30,
+            left: 16,
+            right: 16,
+            child: SizedBox(
+              width: double.infinity,
+              height: 48,
+              child: ElevatedButton(
+                onPressed: () {
+                  Navigator.of(context).push(
+                    MaterialPageRoute(
+                      builder: (context) => const ThirdUploadPage(),
+                    ),
+                  );
+                },
+                style: ElevatedButton.styleFrom(
+                  backgroundColor: Colors.orange,
+                  shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(12),
+                  ),
+                ),
+                child: const Text(
+                  'Next',
+                  style: TextStyle(
+                    fontWeight: FontWeight.bold,
+                    fontSize: 20,
+                    color: Colors.white,
+                  ),
+                ),
+              ),
+            ),
+          ),
+        ],
       ),
     );
   }
